Commit that passes tests on a branch but will fail when merged
This is a commit that was authored on a version of Master before https://github.com/DFE-Digital/apply-for-postgraduate-teacher-training/pull/939/files was merged, and modifies something that will be tested by that branch. The tests on this branch should be green, and it does not produce a merge conflict, but if this is merged, it should blow up!