Move candidate components into namespace
We currently have a large number of candidate-specific components in
the main namespace, and a few in specific namespaces.

This cleanly separates components that can be reused across namespaces
from the ones that are specific.

Not all “review” components could be moved into the Candidate
Interface, because they’re used in the Provider interface too - the
VolunteeringReviewComponent and WorkHistoryReviewComponent. These
should be refactored to be either candidate-specific or more generic.