Update GenerateTestApplications to set edit by date to now
After an application has been submitted, a candidate has 5 working
days to edit their application. This is calculated at submission and
used to work out whether an application is in amend period.

However, when we generate test applications we are able to skip this
period to send applications to a provider as the service we use
does not check the `edit_by` date.

Therefore, it is necessary to update the `edit_by` to now, to
make it come up as no longer amendable.