Generate a state chart minus Provider-invisible states for API docs
Relies on an implementation detail of Workflow, which is the
`workflow_spec` method used to generate the list of states for the
diagram.

Create a disposable class which wraps this method on
ApplicationStateChange so it can pare down the state list.

Didn't create a proper class in a proper file because
- this is a very specific hack with virtually nonexistent potential for reuse
- it's not in any production code paths