Introduce HeaderComponent
This component brings together the logic for constructing header
navigation for the various parts of the application.

Instead of putting links into partials, header navigations are now
configured by NavigationItems.for_candidate_interface,
.for_provider_interface etc.

These methods return lists of NavigationItems which the HeaderComponent
knows how to turn into links.

This means that no logic about what should be in navigations lives in
templates. It exposes some wrinkles about the way we handle eg
SupportUsers vs DfE Sign-in users in the same way across Support and
Provider interfaces and should make refactoring these easier.

This resolves a bug where the dropdown menu was appearing on a screen
where it shouldn't, and hopefully prevents that kind of bug in the
future.

We also get to throw away a lot of repeated markup which risked
being brittle.