Keep track of the feedback status of a reference
We currently have 3 implicit states for a reference. When the
application hasn’t been submitted, the references haven’t been
requested yet. When the candidate submits the application, we request
the references. And after the referees have come back with their
feedback, the feedback is provided.

This commit adds an explicit `feedback_status` to the application
reference, so we can more easily track and filter.

In the future we can:

- Add a state to indicate that a chaser email has been sent
- Add a state to indicate that the user has refused to give a reference
- Add a state to indicate that the email address was invalid