Don't retry syncing from find as much
This limits the number of retries of the sync from find process to 3,
which means that it’ll retry for something like a few minutes
(Sidekiq’s retry mechanism is partly random -
https://github.com/mperham/sidekiq/wiki/Error-Handling#automatic-job-ret
ry). This is okay since we reschedule this job every 15 minutes anyway.

Not having this limit will mean that jobs will keep piling up if Find
ever goes down (not that it ever would, like yesterday).