Include accredited courses on support/provider page
Note whether providers vary, and decide whether or not to show the
provider column if not.

Accredited courses are presented in a separate table so there's no
confusion around the count we display at the top of the page. Add
"Offers" and "Accredits" to the respective table headings so it's clear
which is which.

I considered putting the provider and course into the same column, but
two links in one column felt uncomfortable. Leaving the provider as
plain text was almost equally weird. A new column seemed the best
compromise, but I felt it didn't merit a whole new component. Hence
this fudge.