Don't create associations for completed applications unless necessary
We currently create a lot of references, qualifications and other
associated records when creating a “completed_application_form”.

This commit changes that factory to require an explicit opt-in for
creating the references. This means that a call to
`create(:completed_application_form)` now creates a lot fewer records,
which speeds up the tests.