Test applications don’t have multiple identical course choices
This defect was introduced in #1160 when we turned a loop that called
CourseOption.all.sample into a call to FactoryBot.create_list.

Also, instead of sampling a random CourseOption when we create the
application — which may result in the candidate applying to the same
course at two different sites, not permitted via the UI — ensure we pick
distinct courses. If there aren't enough distinct courses, throw an
exception.