Set feedback_status synchronously when sending reference requests
When this value is set asynchronously, it's possible for the sidekiq job
to complete after the reference has been returned — not a problem IRL,
but it is in sandbox, where references can be instacompleted, or in
GenerateTestApplications.

Rename ApplicationReference#update_token! to refresh_feedback_token!,
which is more explicit, and stop it updating the feedback status as a
side effect.

Callers (ie services) are now responsible for synchronously
setting the feedback_status of references, but as services
are responsible for managing all our status transitions this feels like
the right place to keep this responsibility.