Don't include candidates without forms in the process stats
We currently make the initial query for the process state states
against the candidates table, but then count the number of distinct
states per application form. This means we’re counting the 10 user who
don’t have an application form (in practice, this gets condensed to a
off-by-one error because all form_ids are nil).