Fix bug that causes candidates to apply to a different course
Course codes are unique for a provider, but not unique globally. In the
candidate interface, there are 2 places where we use the course code as
a unique identifier. In cases where a course code is shared between 2
providers, this can cause the candidate to accidentally choose the
wrong course.

This fixes the issue by scoping the course finding to the provider.