Speed up system tests
I was investigating the performance of our system tests when I noticed
a large number of log lines like this:

```
[Webpacker] Everything's up-to-date. Nothing to do
```

I think this is Webpacker trying to compile the assets and then telling
us that no CSS/JS changes were detected.

Turns out this system is *super* expensive. By turning the `compile`
flag off in test causes a preposterous speedup in the tests:

```
Before: Finished in 2 minutes 36.5 seconds
After: Finished in 57.61 seconds
```

Because `compile: true` does a compile on each request, we have to make
sure that our assets are compiled before running the tests, similarly
to how production works. We can do this by sprinkling the compile task
in relevant places:

```
RAILS_ENV=test rails webpacker:compile
```

I’ve also added a reminder that shows every test run, so that if you
run `bundle exec rspec` you won’t get caught out. I think this is a
reasonable trade-off for the speed up that we’re seeing.

Note that there’s something I don’t understand in the compilation - the
current compilation saves the result in `public/packs-test`, but
running our rake task saves it in `public/packs`. But it still works.