Don't swallow invalid transitions in ReceiveReference
We currently add errors to the ReceiveReference object if the
transition fails. This can’t work because the class isn’t an
ActiveModel. This removes the rescue and makes sure that we just raise
the error. Catching the transition error is pointless anyway because
the user won’t know what to do with the error message.