Use different components for references in support & provider UI
We currently show the “Given consent for research?” value to the
provider, which we shouldn’t.

This PR solves the problem by separating the components entirely, and
making the changes to the provider side. I’ve also removed the status
from provider UI, which should always be “feedback provided” (otherwise
providers can’t see it). Also removed a bunch of if-statements that are
no longer necessary, such as the title and whether or not to show the
send email links.