Clarify reference example
Giving the headteacher of where you went to school as a reference wouldn't be appropriate. An old headteacher is not a good reference. A provider suggested it implies that no one from their adult career/studies wants to give a reference.

A headteacher at a place you have volunteered or worked as an adult is a good reference.