Extract the days after which to chase referees into TimeLimitConfig
This is consistent with other places where we do this. I hope we’ll
refactor
this to be nicer at some point.

To make this change we need to be able to calculate a day in the
future, instead of just the past. This in turn requires a bunch of
refactoring.