Don't accidentally create providers when using build_stubbed
We’re seeing providers appear on QA, caused by the mailer previews.
This is because even though we use `build_stubbed`, we override the
initializer and do a `find_or_create_by`.

Also only create a provider agreement when _creating_ a provider,
because otherwise a provider & agreement will be created anyway.