Move service performance dashboard into integrations
This moves the dashboard to /integrations, to make it clearer it’s a
special page that is open to the public. The service performance page
now just links to it. This will make it easier to add non-public stuff
to the support section, and prevents us making things public
accidentally.