Use database IDs instead of Course and Provider codes
In the candidate interface, there are several places where we use codes
to identify Courses and Providers. Certain URL paths contain these
codes, they subsequently end up in the params for the related actions,
and we end up using them in database queries.

The issue with this is that course codes are unique for a provider, but
not unique globally. In cases where a course code is shared between two
providers, this can cause the system to accidentally choose the wrong
course.

Change the route definitions for the CourseChoicesController to use ids
instead of codes. Subsequent commits will deal with the necessary code
changes.