Don't try to look up dbd_date on a choice that might not have an offer
It's possible for an application's first choice not to have a dbd_date
on it, but the mailer assumes it does. Look up the date from one of the
choices in "offer" state.

When making an offer-less application choice in the spec, make sure we
don't set a dbd_date, which doesn't make sense at this point.