Fix for applications export
The application timings CSV is supposed to show the latest time a
column was updated on an application form. However, the method looks at
when the column was last updated including the time the record was
created. Since all columns are present in the changeset of a create, we
basically fall back to the created_at of a application form, making it
look like all columns were updated.