Call rejection email outside of transaction
We think there is a race condition that occurs when we schedule an email for delivery inside this transaction, where the delivery occurs before the transaction has time to finalise and commit the changes to the database. This means that `decline_by_default_at` can be set to nil, which lead to a sentry error in production.