Group course selection system specs
Some of our system specs have a fair amount of crossover and repeat
testing. In order to start consolidating and/or removing unnecessary
tests, it's useful to group together specs which test similar
functionality.

One such area is course selection. Group specs that cover this part of
the codebase into a folder, as a first step in the process of
refactoring them.