Make erb-lint work with Rubocop 0.80
Since 0.78 rubocop has thrown an exception on a missing cop. Because of
the funky way erblint loads the rubocop config it ends up not correctly
resolving the dependency on rubocop-rails. The solution (until someone fixes
it) is to explicitly require rubocop-rails at the top of
rubocop_todo.yml.

https://github.com/openstreetmap/openstreetmap-website/issues/2472#issuecomment-568255229

Fix a deprecation warning in the erblint rubocop config.