Prepend Heroku app name to config.hosts
We call `authorised_hosts.first` in `HostingEnvironment#hostname` when
`ENV["CUSTOM_HOSTNAME"]` is not defined. On Heroku, `authorised_hosts`
always contains `apply-for-teacher-training.herokuapp.com` as the first
item in a review app. Because of this, our review apps send magic link
emails that tell users to go to `apply-for-teacher-training.herokuapp.com`.

This is a bit annoying, but we can easily work around it by prepending
the desired hostname.