If there is only one filter in a group it will not render
Ensures that if there is only one filter in a group then it will not
render as a filter.

It does not make sense to render a filter of there only one of a
particular type available as it would return the who set of results.

For instance if there is only one provider available as a filter then
all of the applications for a provider_user will belong to that one
provider, so filtering by it would return all results in any case.