Freeze time so that tests pass correctly.
These tests passed on friday 13th of March but not today (16th)

Why?

Reason being that we set 'decline_by_default` in the spec to be
ten business days in the future. Then check this in the email and expect
it to be the the same day.

This fails if you are less than ten business days away from the clocks
going forward, because when we set the declined by default date using
TimeLimitCalculator it sets the date to be at the end of the day, and
this is pushed over to the next day when the clocks go forward.