Fix bug when users double submit the find course selection form
If a user comes in from find, adds a course to their application that
brings them up to 3 courses, and then goes back with their browser and
resubmits the form, they'll get a 500 because :options_for_site is not
defined in this controller.

Instead, we should display an error message and direct them to the
index page for course choices.