Always execute Sidekiq jobs in system tests
System tests almost always need Sidekiq to run. In some places we've
turned this on by tagging `sidekiq: true`. In cases where we send
emails in the background we needed a slightly different helper
`perform_enqueued_jobs` (this is part of ActiveJob).

Now that all emails run in the background, it's easier to always turn
on Sidekiq/email processing in system tests.