Use the SetDeclineByDefault service when recalculating
This changes the date recalculation worker to use the existing service
to make the changes.

This fixes a potential issue with the logic - the recalculation worker
does not look up the correct effective_date like `SetDeclineByDefault`
does.

Also simplifies the work the recalculation worker has to do, so the
tests can become simpler as well.