Add invalidated_by_find field to CourseOptions
When syncing Find data in SyncProviderFromFind, we create or update the
course options for each course. Pre-existing course options can become
invalid if the course is no longer available at a given site. The plan
is to delete these course options if they aren't associated with an
ApplicationChoice. If they are, it would be useful to mark them in some
way. The intention is to set this field to true during the sync if a
course option becomes invalid.