Validate presence of first_ and last_ name when creating a ProviderUser
The API call to DfE Sign in will end up validating these properties, but
there's no guarantee it always will, and we depend on them. Also, the
errors returned by DSI are not helpful to the user as they refer to
fields in the API call, not in the form.