Change how study modes are determined in SyncProviderFromFind
When we upsert course options during the sync, part of the process
involves iterating over available study modes.

Determining available study modes is currently subject to the following
issue:

- We pull a Course (C) with full_time_or_part_time study_mode.
- We create CourseOptions (O1 and O2) with full_time and part_time
  study_modes respectively.
- The study_mode of C changes to full_time because there are no longer
  any part time places.
- On next sync, we use the study_mode of C to find the associated
  CourseOptions. We find O2 because it's part time, but don't find O1.
  Therefore we don't update O1's vacancy status, which is bad because it
  remains available for applications when it's actually full.

Fix this by introducing an object that helps determine the appropriate
study modes list. It returns either one or both study modes depending on
the current and historical study mode of the course.