Sync: refactor creation of Course Options
We have a validation to ensure the presence of `vacancy_status`. This
means that `CourseOption.find_or_create_by` actually always fails.

Unfortunately, this doesn't raise an error in Rails but just results in
a CourseOption that isn't persisted.

Luckily, we fix our own bug in the last line of the iteration, where we
`update` the object and it's persisted after all (however, if something
else is wrong then we silently fail because it'll return true/false).