Require users to be signed in for the interstitial route
An error occurred on production that is caused by a non-signed-in
candidate visiting the /candidate/interstitial URL. I'm assuming that
they pressed the back button or something similar.

https://sentry.io/organizations/dfe-bat/issues/1534179482

This solves the issue by pulling out the interstitial method into a
separate controller, which doesn't skip `authenticate_candidate!` and
will require candidates to be signed in. Bonus that it cleans up the
original controller, which is pretty large already.