Preserve Find course params through sign in
Current system behaviour is for Apply From Find to send the user to the
eligibility screen (typically seen when signing up to the service for
the first time). Pre-existing candidates see the same flow.

To sign in to their accounts they have to click a link in the header or
on the start page. In doing so, their Find course params aren't
preserved across the sign in process. Furthermore, now that we've added
a 'create account or sign up' page, this issue is more pronounced as we
present an explicit option to sign in after the Apply From Find screen.

This commit fixes the problem by preserving course selection params
across the sign in process.