Present study mode selection after 'you_selected_a_course_page'
The 'you_selected_a_course_page' feature handles Find course selections
for authenticated users. When confirming the course selection, it
redirects the candidate to pick a site. This causes a bug if the course
study_mode is 'both_full_time_and_part_time', as the site selection page
relies on one or the other study mode being picked to filter the site
list correctly.

Fix this by sending the candidate to the study mode selection screen
first if there is a choice of study modes.