Show the inline error under the heading for an invalid provider response
When a provider responds to an application, they need to choose one of
the four options. If they don’t choose the correct one, then they need
to see an error.

The legend element on this checkbox fieldset does double service as the
<legend> element itself and the <h1> for the whole page. This makes
things a bit difficult for the form builder, as it assumes that anything
inside the `govuk_radio_buttons_fieldset` block is part of the field
itself. Having the legend as part of the block causes the little red
inline error message to appear above the heading, next to the top error
where it's useless.

The design system prefers that we put the inline error underneath the
legend. (https://design-system.service.gov.uk/components/error-summary/).

The form builder does support passing a `legend` option to the
fieldset, which DOES behave correctly, but this doesn't work for us
because we want custom markup (the caption containing the user's name)
in the legend.

As we can't do that, cheat: pull the <legend> out of the fieldset so
that the error always appears underneath it.

It's not ideal because semantically the <legend> belongs inside the
radio buttons' fieldset, and the legend itself doesn't receive a red
border when the field contains an error. But it's better than it was.