Authenticate sign ins from 'create account or sign in' page
On the 'create account or sign in' page, if a candidate selects the
latter they currently get sent to the dedicated sign in page.

In order to make this slightly simpler for the candidate, change it so
that we ask for an email and begin the magic link authentication process
immediately without the additional step.

- Add an email field to the 'create account or sign in' page and its
  corresponding form.
- Extract existing sign in logic into a service object.
- Call this service from both the SignInController and
  StartPageController.