Don't allow withdrawal for applications not sent to provider
We currently allow applications to be withdrawn by candidates in the
"application_complete" state (when references have been received but
the application is still in the 5 day edit window).

We've realised that withdrawing an application that hasn't been sent to
the provider causes a number of undesirable side effects:

- The application becomes visible to the provider because the
`withdrawn` state is not hidden to providers
- We send out emails about the withdrawal to providers. This is
confusing because the provider wouldn't have seen the application
before the withdrawal.