Add selectable scope to CourseOption model
When we sync with Find, occasionally some course options may be get
marked as invalidated_by_find if the site they represent is no longer a
valid choice for that course.

In the majority of cases, these course options will be deleted. A
minority, however, may have been chosen by a candidate at some point and
are therefore associated with their application. We don't delete those,
and instead follow a manual process to fix them.

In order to prevent any other candidates selecting these invalidated
options, introduce a scope to the CourseOption model that filters them
out. Use this scope in the site selection form.