Add system spec for unavailable course option warnings
To help ensure candidates are applying only to courses that are
available and have vacancies, we want to display warnings on the
application review page if a course option falls into one of several
states that makes them unavailable.

Add a system spec outlining this behaviour, including descriptions of
the specific states we look for.