Add the export of applications data for UCAS to the support interface
To support the dual-running of Apply and UCAS until Apply eventually takes over completely,
we need to detect applications which have been duplicated between the two systems.

To do this, we've agreed to send UCAS an export of all submitted-or-later applications which
they can match against their system - they'll then send us another file back indicating any
which have been matched, so that we can take corrective action.

This PR implements just the first part of this - a way to export the datafile as CSV in a
format that UCAS have specified. The MVP version is to just add this as a manual download
in the support interface. Once the format is finally agreed, we can (separately) automate
the export and transfer to UCAS in a job.

For further details, see the trello card #388 on the Candidate Team board