Fix issue with govuk_check_box checked value
The underlying issue is that ActionView::Helpers::FormBuilder looks up the
value to determine 'checked' state from a method name on the object.
Our ProviderUserForm does not respond to 'manage_users' so this fails to
populate the checkbox. Hence the ugly use of check_box_tag until we find
a better way to do this. It could be that the structure of the data on the
form is wrong, nesting values in a hash certainly causes problems.
One solution might be to nest a class which encapsulates permissions with
methods instead of hash data.