Make form permissions work with govuk_check_box
The previous commit used check_box_tag to solve the issue of replaying
permissions back to the form check boxes and set the checked attribute.
The underlying issue was how action view tag helpers determine the
value of the input and use this to determine the checked state.
By using form_for with form permissions as the second argument and
passing back permissions as an OpenStruct the check box tag can
call the method 'manage_users' on permissions and work out the
checked / unchecked state.