Use SaveProviderUser service for user updates
In order to deactivate permissions we need to unset
any permissions not sent via the edit provider user form.
It's simpler to unset everything and then set the values
from the form. This is wrapped in a transaction.
This is handled by the SaveProviderUser service, this
commit amends the controller update action to use the new
service.