Introduce the basis for changing courses in support
Currently we have a feature in support to add a course to an
application that has been submitted. We want to build a feature to
allow support to also remove courses from someone's application.

This is trickier than it sounds at first. Removing a course means
different things depending on how many courses the candidate has
applied to:

1. If the candidate has applied to a single course, removing a course
will end the application. They will have to apply again via Apply 2
(currently only available on UCAS).
2. If the candidate has applied to a single course but wants to change
it to another, we would want add a new course first, and then remove
the original one to avoid scenario #1 from kicking in.
3. If the candidate has applied to multiple courses, we can just remove
the course from the application without any side effects.

To implement the feature we'll implement a flow that allows support
users to make specific changes:

1. Add a course (only available when 1 or 2 choices have been made)
2. Replace a course (always available)
3. Withdraw from a course (only available when 2 or 3 choices have been
made)
4. Withdraw from all courses and cancel application (always available)

This will allow us to easily implement the changes and handle the side
effects like email notifications, Apply 2, and cancelling referees.

This PR takes the existing feature and moves it into the flow, so we
can build the individual features on top of it.