Send application_submitted provider email in after_commit callback
In several of our environments, including production, we see the
following exception:

    ActionView::Template::Error
    wrong number of arguments (given 1, expected 0)

It appears intermittently and occurs when trying to send the
ProviderMailer#application_submitted email. Its source is this line
in the email template:

    @application.rbd_date.to_s(:govuk_date)

The error occurs because the `rbd_date` is nil (this is just a struct
attribute that holds the `reject_by_default_at` value of the
application_choice record). When inspecting a sample of the application
choice records that are subject to this error, the
`reject_by_default_at` value appears to be correctly populated.

The email is sent as part of the SendApplicationToProvider service
object, which updates the `reject_by_default_at` just before doing an async
send of the email. The error above is possibly due to a worker
attempting to send the email before the database commit that stores the
RBD value has finished.

Attempt to fix this by sending the email in an after_commit callback on
the ApplicationChoice model. Pass in a temporary flag on the model so
that the email is only sent in this particular scenario.