Remove transaction block in SubmitReference
In several of our environments, including production, we see the
following exception:

    ActionView::Template::Error
    wrong number of arguments (given 1, expected 0)

It appears intermittently and occurs when trying to send the
ProviderMailer#application_submitted email. Its source is this line
in the email template:

    @application.rbd_date.to_s(:govuk_date)

The error occurs because the rbd_date is nil (this is just a struct
attribute that holds the reject_by_default_at value of the
application_choice record).

The likely cause of this seems to be the transaction nesting that occurs
when SubmitReference calls the SendApplicationsToProvider service. The
call to this service is part of the top-level transaction. It iterates
over eligible application forms and calls SendApplicationToProvider for
each of their application choices. This service then contains database
changes wrapped in another transaction, followed by async delivery of
the application_submitted email.

Because these database changes are only committed when the top-level
transaction block closes, it seems likely that on some occasions a
Sidekiq worker is attempting to send the email before the
reject_by_default_at is present in the database record.

This commit removes the transaction nesting, and instead leaves a linear
series of database updates and transactions for each discrete step of
the SubmitReference process.