Remove edit application feature
This feature was never turned on in production, for reasons outlined in
the design history:

https://bat-design-history.netlify.app/apply-for-teacher-training/amendi
ng-mvp

This removes the feature altogether. It renames the main before_action
to "redirect_to_dashboard_if_submitted", which is what it was before
the feature was introduced. It also adds it to some missing controllers
(it was easy to see here which controllers didn't have changes made to
them when doing a project search-replace).