Prevent candidates from being able to add a course twice via sites
We validate for this as part of `PickCourseForm`, which works fine
for courses without multiple sites. For courses with multiple sites,
candidates can get around this by navigating to the sites page directly
and submitting it.

This adds checks to the site options page as well as the POST action,
and redirects the user to the courses review page with an appropriate
error flash.