Refactor SubmitReference
In addition to making the business logic slightly clearer, this change
simplifies the database changes that occur when enough references have
been submitted. In this scenario, all record updates handled by
SubmitReference occur within a single transaction, allowing them to
roll back to a consistent state in the event of any errors.