Mark emails that we can't send to GOV.UK Notify
This marks emails in the email log with the delivery status
`notify_error` if we encounter an error sending the email to GOV.UK
Notify. It comes on the back of the recent PaaS incident that took down
the service for a couple of hours.

Note that we `rescue_from` all Notify exceptions. I originally tried to
distinguish between client errors like invalid API key or email address
(anything descendent from `Notifications::Client::ClientError` and
server errors (`Notifications::Client::ServerError`). The idea was that
we could have separate statuses for each, and perhaps skip the retries
for client errors. However, this is not going to work because the way
GOV.UK Notify system behaved during the incident.

We did see some Bad Gateway errors, which raised a ServerError
(https://sentry.io/organizations/dfe-bat/issues/1568629644). But other
errors were 404
(https://sentry.io/organizations/dfe-bat/issues/1623747428) and a 400
(https://sentry.io/organizations/dfe-bat/issues/1569990518). This means
we can't reliably distinguish between client and server errors from
Notify, and we'll have to treat them the same.