Make the Email#notify_reference unique
This appends a random string to the notify reference for emails with a
stable reference. We need this because when an email fails to send
(GOV.UK Notify returns an error), we'll start retrying. This causes
multiple emails in the log with the same `notify_reference`.

When getting the callback from Notify it becomes tricky to determine
which of the emails the delivery receipt is for.