Upgrade to ViewComponent
This is the new name for ActionView::Component

The initialization code is slightly different, but the semantics are the
same. ViewComponent removed ActiveModel::Validations from the ::Base
class from which all components inherit - we have a lot of that in our
codebase. While we decide what to do about it, monkey-patch
ViewComponent::Base :D

Include the ViewComponent::TestHelpers using a general metadata rule.