Update candidate_sign_in_url helper to return magic link
This helper is used by several CandidateMailer emails. Currently it
returns the URL for the sign in page, with an encrypted candidate id in
the params. This simply redirects candidates to an 'expired token' page
from where they have to click a button to get a magic link.

Improve this experience by changing the helper to return a magic link
instead, ie - an authentication endpoint with the encrypted candidate id
and a sign in token in the params. This allows the candidate to
authenticate immediately if they click the link before token expiry.
It's then only after token expiry that they see the 'expired token'
page.