Remove expired sign in branch on candidate SignInController#new
The only codepath that led down this branch appears to have been the
magic link helper used in several CandidateMailer emails. That helper
has now been updated to return an authentication link instead, making
this if statement unnecessary.

Remove the params check, leaving the rendering of the sign in page the
sole responsibility of this action.