Add UCAS 1st of June recalculation to RecalculateDates worker
To ensure that RBD/DBD dates are consistent with UCAS approach, we need
to update our worker.

All applications that had an RBD or DBD that fell between the 20th of
April and 31st of May *before* we made that period a working day freeze
should instead be recalculated for the end of the 1st of June.