Switch codebase to use CourseOption#site_still_valid
We're renaming invalidated_by_find to site_still_valid so it's clearer
what the purpose of this field is.

To avoid errors in the frontend and the Find API sync, deploy in three
steps.

1. Add a column with the new name and ensure we write to both when
   updating a record.
2. Migrate old data to new column, and introduce code changes that use
   the new column.
3. Remove old column.

This commit is step 2. It also includes code to ignore
invalidated_by_find from the CourseOption column list, ensuring it
doesn't appear in cached schema definitions.