Restrict course selection when candidates Apply Again
When a candidate's application is unsuccessful, they can opt to Apply
Again. One of the business rules for the Apply Again process is that
they can only apply to one course at a time.

Add a passing implementation of this behaviour, containing UI changes to
restrict selection of additional courses for application forms in this
state.