Adds Accredited body to applications show
Not all courses are accredited by different providers than
the course provider. In these cases the corse provider is self
accrediting and so their name should appear in the Accredited
Provider field on the Course information section of the show
view.

Added a new CourseDetailsComponent to encapsulate this
logic also added new spec it.