Change how PickSiteForm validation error is handled
Currently, when the CourseChoiceController#pick_site_for_course method
tries to save the PickSiteForm, a vlidation error results in a render of
the options_for_site template. This leads to a slightly buggy rendering
of that page with no sites listed (because we don't set the expected
instance variable containing the relevant sites).

Change this so that we populate the flash with the first validation
error reported by PickSiteForm, and redirect to the candidate's
application form.