Remove some of the 'apply_again' feature flag checks
We don't want to check for the presence of this flag in all cases. While
we should check for it around the code that actually allows a candidate
to Apply Again, once an application is already in that phase it would be
preferable for them to continue seeing the apply_again UI whether or not
the feature is enabled.

This prevents scenarios where, eg - they can suddenly add 3 courses
rather than 1 because the feature is switched off.