Prevent 500 on Application PDFs
Move Rack::Deflater to the top of the middleware stack as recommended in
https://www.schneems.com/2017/11/08/80-smaller-rails-footprint-with-rack-deflate/

We added the Rack::Deflater middleware to compress our responses. This
middleware returns a GZip stream rather than a string.

Unfortunately PDFKit ended up next to it in the chain, and tried to
turn the gzip stream into a PDF. This caused an error here:

https://github.com/pdfkit/pdfkit/blob/master/lib/pdfkit/middleware.rb#L23

Exactly why this happened and who's in the wrong here is beyond scope
for this commit.